Cache model methods after compilation
